refactor Configuration selection test
